added perplexity, ratings

+++ remi_evaluate.py
@@ -17,6 +17,7 @@ from os.path import exists, basename, isdir
 from os import mkdir, listdir
 from typing import Union, List
 import multiprocessing
+import math
 
 import numpy as np
 import pandas as pd
@@ -25,9 +26,10 @@ import torch.utils.data
 from tqdm import tqdm
 import x_transformers
 
-from dataset_full import pitch_class_entropy, scale_consistency, groove_consistency, MMT_STATISTIC_COLUMNS, CHUNK_SIZE
+import dataset_full
 import remi_representation
-from remi_train import OUTPUT_DIR, BATCH_SIZE
+import remi_dataset
+import remi_train
 import utils
 
 ##################################################
@@ -37,7 +39,7 @@ import utils
 ##################################################
 
 # default number of samples to evaluate
-N_SAMPLES = 1000
+N_SAMPLES = remi_train.BATCH_SIZE * 128
 
 # evaluation constants
 SEQ_LEN = 1024
@@ -45,7 +47,16 @@ TEMPERATURE = 1.0
 FILTER = "top_k"
 
 # output columns
-OUTPUT_COLUMNS = ["model", "path"] + MMT_STATISTIC_COLUMNS
+OUTPUT_COLUMNS = ["model", "path"] + dataset_full.MMT_STATISTIC_COLUMNS + ["tracks", "loss"]
+
+##################################################
+
+
+# HELPER FUNCTIONS
+##################################################
+
+# perplexity function
+perplexity_function = lambda loss: math.exp(-math.log(loss))
 
 ##################################################
 
@@ -56,14 +67,12 @@ OUTPUT_COLUMNS = ["model", "path"] + MMT_STATISTIC_COLUMNS
 def parse_args(args = None, namespace = None):
     """Parse command-line arguments."""
     parser = argparse.ArgumentParser(prog = "Evaluate", description = "Evaluate a REMI-Style Model.")
-    parser.add_argument("-i", "--input_dir", default = OUTPUT_DIR, type = str, help = "Dataset facet directory containing the model(s) (as subdirectories) to evaluate")
-    parser.add_argument("-ns", "--n_samples", default = N_SAMPLES, type = int, help = "Number of samples to evaluate")
+    parser.add_argument("-i", "--input_dir", default = remi_train.OUTPUT_DIR, type = str, help = "Dataset facet directory containing the model(s) (as subdirectories) to evaluate")
     # model
     parser.add_argument("--seq_len", default = SEQ_LEN, type = int, help = "Sequence length to generate")
     parser.add_argument("--temperature", nargs = "+", default = TEMPERATURE, type = float, help = f"Sampling temperature (default: {TEMPERATURE})")
     parser.add_argument("--filter", nargs = "+", default = FILTER, type = str, help = f"Sampling filter (default: '{FILTER}')")
     # others
-    parser.add_argument("-bs", "--batch_size", default = BATCH_SIZE, type = int, help = "Batch size")
     parser.add_argument("-g", "--gpu", default = -1, type = int, help = "GPU number")
     parser.add_argument("-j", "--jobs", default = int(multiprocessing.cpu_count() / 4), type = int, help = "Number of workers for data loading")
     parser.add_argument("-r", "--reset", action = "store_true", help = "Whether or not to regenerate samples")
@@ -84,7 +93,7 @@ if __name__ == "__main__":
     args = parse_args()
 
     # get directories to eval
-    model_dirs = list(filter(lambda path: isdir(path) and path.endswith("M"), map(lambda base: f"{args.input_dir}/{base}", listdir(args.input_dir))))
+    model_dirs = list(filter(lambda path: isdir(path) and path.split("_")[0].endswith("M"), map(lambda base: f"{args.input_dir}/{base}", listdir(args.input_dir))))
     models = list(map(basename, model_dirs))
 
     # set up the logger
@@ -97,6 +106,9 @@ if __name__ == "__main__":
     logging.info(f"Saved arguments to {args_output_filepath}")
     utils.save_args(filepath = args_output_filepath, args = args)
     del args_output_filepath
+
+    # data paths filepath
+    data_paths_filepath = f"{args.input_dir}/" + ("test" if exists(f"{args.input_dir}/test.txt") else "valid") + ".txt"
 
     # get the specified device
     device = torch.device(f"cuda:{abs(args.gpu)}" if (torch.cuda.is_available() and args.gpu != -1) else "cpu")
@@ -126,18 +138,10 @@ if __name__ == "__main__":
     # output file
     output_filepath = f"{args.input_dir}/evaluation.csv"
     output_columns_must_be_written = (not exists(output_filepath)) or args.reset
-    n_lines_in_output = sum(1 for _ in open(output_filepath, "r")) if exists(output_filepath) else 0
-    if output_columns_must_be_written or (n_lines_in_output == 1): # if column names need to be written
+    n_batches = int((N_SAMPLES - 1) / remi_train.BATCH_SIZE) + 1
+    if output_columns_must_be_written: # if column names need to be written
         pd.DataFrame(columns = OUTPUT_COLUMNS).to_csv(path_or_buf = output_filepath, sep = ",", na_rep = utils.NA_STRING, header = True, index = False, mode = "w")
-        n_samples_to_calculate = utils.rep(x = args.n_samples, times = len(model_dirs)) # number of samples to calculate
-        starting_indicies = utils.rep(x = 0, times = len(model_dirs))
-    else:
-        previous = pd.read_csv(filepath_or_buffer = output_filepath, sep = ",", na_values = utils.NA_STRING, header = 0, index_col = False) # load in previous values
-        n_samples_to_calculate = previous.groupby(by = "model").size().to_dict() # calculate number of samples already generated
-        n_samples_to_calculate = list(map(lambda model: max(args.n_samples - n_samples_to_calculate.get(model, 0), 0), models)) # determine number of samples to calculate
-        starting_indicies = list(map(lambda model: max(list(map(lambda path: int(path[len(f"{args.input_dir}/{model}/eval/"):-len(f".npy")]), previous[previous["model"] == model]["path"])) + [-1]) + 1, models)) # get the starting index for generation names for each model
-        del previous
-    del output_columns_must_be_written, n_lines_in_output # free up memory
+    del output_columns_must_be_written # free up memory
 
     ##################################################
 
@@ -154,9 +158,14 @@ if __name__ == "__main__":
 
         # return a dictionary
         if len(music.tracks) == 0:
-            return utils.rep(x = np.nan, times = len(MMT_STATISTIC_COLUMNS))
+            return utils.rep(x = np.nan, times = len(dataset_full.MMT_STATISTIC_COLUMNS)) + [""]
         else:
-            return [pitch_class_entropy(music = music), scale_consistency(music = music), groove_consistency(music = music)]
+            return [
+                dataset_full.pitch_class_entropy(music = music),
+                dataset_full.scale_consistency(music = music),
+                dataset_full.groove_consistency(music = music),
+                dataset_full.get_tracks_string(tracks = music.tracks),
+            ]
 
     ##################################################
 
@@ -164,7 +173,7 @@ if __name__ == "__main__":
     # REPEAT WITH EACH MODEL IN INPUT DIRECTORY
     ##################################################
 
-    for model_name, model_dir, n_samples, starting_index in zip(models, model_dirs, n_samples_to_calculate, starting_indicies):
+    for model_name, model_dir in zip(models, model_dirs):
 
         # LOAD MODEL
         ##################################################
@@ -177,7 +186,26 @@ if __name__ == "__main__":
         # load training configurations
         train_args_filepath = f"{model_dir}/train_args.json"
         train_args = utils.load_json(filepath = train_args_filepath)
-        del train_args_filepath       
+        del train_args_filepath
+
+        # load dataset and data loader
+        dataset = remi_dataset.MusicDataset(
+            paths = data_paths_filepath,
+            encoding = encoding,
+            indexer = indexer,
+            encode_fn = remi_representation.encode_notes,
+            max_seq_len = train_args["max_seq_len"],
+            max_beat = train_args["max_beat"],
+            use_augmentation = train_args["aug"],
+        )
+        data_loader = torch.utils.data.DataLoader(
+            dataset = dataset,
+            num_workers = args.jobs,
+            collate_fn = dataset.collate,
+            batch_size = remi_train.BATCH_SIZE,
+            shuffle = False
+        )
+        data_iter = iter(data_loader)
 
         # create the model
         model = x_transformers.TransformerWrapper(
@@ -210,39 +238,60 @@ if __name__ == "__main__":
         ##################################################
 
         # iterate over the dataset
-        batch_size_evenly_divides_n_samples = ((n_samples % args.batch_size) == 0) # does batch size evenly divide n_samples
-        n_batches = int(n_samples / args.batch_size) + int(not batch_size_evenly_divides_n_samples) # n batches
         with torch.no_grad():
             for i in tqdm(iterable = range(n_batches), desc = f"Evaluating the {model_name} Model"):
 
                 # get number of samples to calculate
-                n_samples_in_batch = (n_samples % args.batch_size) if ((i == (n_batches - 1)) and (not batch_size_evenly_divides_n_samples)) else args.batch_size
+                n_samples_in_batch = (((N_SAMPLES - 1) % remi_train.BATCH_SIZE) + 1) if (i == (n_batches - 1)) else remi_train.BATCH_SIZE
 
-                # get start tokens
-                prefix = torch.ones(size = (n_samples_in_batch, 1), dtype = torch.long, device = device) * sos
+                # get output filepaths for generated sequences
+                generated_output_filepaths = list(map(lambda j: f"{eval_dir}/{(i * remi_train.BATCH_SIZE) + j}.npy", range(n_samples_in_batch)))
 
-                # generate new samples; unconditioned generation
-                generated = model.generate(
-                    prompts = prefix,
-                    seq_len = args.seq_len,
-                    eos_token = eos,
-                    temperature = args.temperature,
-                    filter_logits_fn = filter_logits_fn,
-                )
-                generated = torch.cat(tensors = (prefix, generated), dim = 1).cpu().numpy()
+                # generate if needed
+                if (not all(map(exists, generated_output_filepaths))) or args.reset:
 
-                # save generation
-                generated_output_filepaths = list(map(lambda j: f"{eval_dir}/{starting_index + j}.npy", range(len(generated))))
-                starting_index += len(generated) # update starting index
-                for j in range(len(generated)):
-                    np.save(file = generated_output_filepaths[j], arr = generated[j]) # save generation to file
+                    # get start tokens
+                    prefix = torch.ones(size = (n_samples_in_batch, 1), dtype = torch.long, device = device) * sos
+
+                    # generate new samples; unconditioned generation
+                    generated = model.generate(
+                        prompts = prefix,
+                        seq_len = args.seq_len,
+                        eos_token = eos,
+                        temperature = args.temperature,
+                        filter_logits_fn = filter_logits_fn,
+                    )
+                    generated = torch.cat(tensors = (prefix, generated), dim = 1).cpu().numpy()
+
+                    # save generation
+                    for j in range(len(generated)):
+                        np.save(file = generated_output_filepaths[j], arr = generated[j]) # save generation to file
+
+                # reload generated files
+                else:
+
+                    # load in generated content
+                    generated = remi_dataset.pad(data = list(map(np.load, generated_output_filepaths)))
 
                 # analyze 
                 with multiprocessing.Pool(processes = args.jobs) as pool:
-                    results = pool.map(func = evaluate, iterable = generated, chunksize = CHUNK_SIZE)
+                    results = pool.map(func = evaluate, iterable = generated, chunksize = dataset_full.CHUNK_SIZE)
+
+                # get loss values for perplexity
+                try:
+                    batch = next(data_iter)
+                except (StopIteration):
+                    data_iter = iter(data_loader) # reinitialize dataset iterator if necessary
+                    batch = next(data_iter)
+                loss_batch = model( # compute loss
+                    x = batch["seq"][:n_samples_in_batch].to(device),
+                    return_outputs = False,
+                    mask = batch["mask"][:n_samples_in_batch].to(device),
+                )
+                loss_batch = float(loss_batch) / n_samples_in_batch # divide by number of samples so we get per sample loss, instead of per batch loss
 
                 # write results to file
-                results = pd.DataFrame(data = map(lambda j: [model_name, generated_output_filepaths[j]] + results[j], range(len(results))), columns = OUTPUT_COLUMNS)
+                results = pd.DataFrame(data = map(lambda j: [model_name, generated_output_filepaths[j]] + results[j] + [loss_batch], range(n_samples_in_batch)), columns = OUTPUT_COLUMNS)
                 results.to_csv(path_or_buf = output_filepath, sep = ",", na_rep = utils.NA_STRING, header = False, index = False, mode = "a")
                 
         ##################################################
@@ -256,11 +305,12 @@ if __name__ == "__main__":
     # log statistics
     bar_width = 104
     results = pd.read_csv(filepath_or_buffer = output_filepath, sep = ",", na_values = utils.NA_STRING, header = 0, index_col = False) # load in previous values
-    for model in sorted(models, key = lambda model: int(model[:-1])):
+    for model in sorted(models, key = lambda model: int(model.split("_")[0][:-1]) + (0.5 if remi_train.FINE_TUNING_SUFFIX in model else 0)):
         results_model = results[results["model"] == model]
         logging.info(f"\n{f' {model} ':=^{bar_width}}")
-        for mmt_statistic in MMT_STATISTIC_COLUMNS:
+        for mmt_statistic in dataset_full.MMT_STATISTIC_COLUMNS:
             logging.info(f"{mmt_statistic.replace('_', ' ').title()}: mean = {np.nanmean(a = results_model[mmt_statistic], axis = 0):.4f}, std = {np.nanstd(a = results_model[mmt_statistic], axis = 0):.4f}")
+    logging.info(f"Perplexity: {perplexity_function(loss = sum(results_model['loss']) / n_batches):.4f}")
     print("\n")
 
     ##################################################
